[query-engine] Refactor StringReplace into text group. (#935)
## Changes

* Introduce a group for scalars which operate on text and move
`ReplaceString` into that group
