+++ rust/experimental/query_engine/engine-recordset/src/scalars/mod.rs
@@ -3,9 +3,11 @@ pub(crate) mod math_scalar_expressions;
 pub(crate) mod parse_scalar_expressions;
 pub(crate) mod scalar_expressions;
 pub(crate) mod temporal_scalar_expressions;
+pub(crate) mod text_scalar_expressions;
 
 pub use convert_scalar_expressions::*;
 pub use math_scalar_expressions::*;
 pub use parse_scalar_expressions::*;
 pub use scalar_expressions::*;
 pub use temporal_scalar_expressions::*;
+pub use text_scalar_expressions::*;
