+++ rust/experimental/query_engine/expressions/src/scalars/mod.rs
@@ -4,6 +4,7 @@ pub(crate) mod parse_scalar_expression;
 pub(crate) mod scalar_expressions;
 pub(crate) mod statics;
 pub(crate) mod temporal_scalar_expression;
+pub(crate) mod text_scalar_expression;
 
 pub use convert_scalar_expression::*;
 pub use math_scalar_expression::*;
@@ -11,3 +12,4 @@ pub use parse_scalar_expression::*;
 pub use scalar_expressions::*;
 pub use statics::*;
 pub use temporal_scalar_expression::*;
+pub use text_scalar_expression::*;
