+++ rust/experimental/query_engine/expressions/src/scalars/scalar_expressions.rs
@@ -44,10 +44,6 @@ pub enum ScalarExpression {
     /// Contains scalar functions for performing parsing operations.
     Parse(ParseScalarExpression),
 
-    /// Returns a string with all occurrences of a lookup value replaced with a
-    /// replacement value or null for invalid input.
-    ReplaceString(ReplaceStringScalarExpression),
-
     /// Returns a slice of characters from an inner string value, a slice of
     /// items from an inner array value, or null for invalid input.
     Slice(SliceScalarExpression),
@@ -57,6 +53,9 @@ pub enum ScalarExpression {
 
     /// Resolve a static value provided directly in a query.
     Static(StaticScalarExpression),
+
+    /// Contains scalar functions for performing text operations.
+    Text(TextScalarExpression),
 
     /// Resolve a value from a query variable.
     ///
@@ -84,10 +83,10 @@ impl ScalarExpression {
             ScalarExpression::Case(c) => c.try_resolve_value_type(pipeline),
             ScalarExpression::Convert(c) => c.try_resolve_value_type(pipeline),
             ScalarExpression::Length(l) => l.try_resolve_value_type(pipeline),
-            ScalarExpression::ReplaceString(r) => r.try_resolve_value_type(pipeline),
             ScalarExpression::Slice(s) => s.try_resolve_value_type(pipeline),
             ScalarExpression::Parse(p) => p.try_resolve_value_type(pipeline),
             ScalarExpression::Temporal(t) => t.try_resolve_value_type(pipeline),
+            ScalarExpression::Text(r) => r.try_resolve_value_type(pipeline),
             ScalarExpression::Math(m) => m.try_resolve_value_type(pipeline),
         }
     }
@@ -112,10 +111,10 @@ impl ScalarExpression {
             ScalarExpression::Case(c) => c.try_resolve_static(pipeline),
             ScalarExpression::Convert(c) => c.try_resolve_static(pipeline),
             ScalarExpression::Length(l) => l.try_resolve_static(pipeline),
-            ScalarExpression::ReplaceString(r) => r.try_resolve_static(pipeline),
             ScalarExpression::Slice(s) => s.try_resolve_static(pipeline),
             ScalarExpression::Parse(p) => p.try_resolve_static(pipeline),
             ScalarExpression::Temporal(t) => t.try_resolve_static(pipeline),
+            ScalarExpression::Text(r) => r.try_resolve_static(pipeline),
             ScalarExpression::Math(m) => m.try_resolve_static(pipeline),
         }
     }
@@ -135,10 +134,10 @@ impl Expression for ScalarExpression {
             ScalarExpression::Case(c) => c.get_query_location(),
             ScalarExpression::Convert(c) => c.get_query_location(),
             ScalarExpression::Length(l) => l.get_query_location(),
-            ScalarExpression::ReplaceString(r) => r.get_query_location(),
             ScalarExpression::Slice(s) => s.get_query_location(),
             ScalarExpression::Parse(p) => p.get_query_location(),
             ScalarExpression::Temporal(t) => t.get_query_location(),
+            ScalarExpression::Text(r) => r.get_query_location(),
             ScalarExpression::Math(m) => m.get_query_location(),
         }
     }
@@ -156,10 +155,10 @@ impl Expression for ScalarExpression {
             ScalarExpression::Constant(c) => c.get_name(),
             ScalarExpression::Convert(c) => c.get_name(),
             ScalarExpression::Length(_) => "ScalarExpression(Length)",
-            ScalarExpression::ReplaceString(_) => "ScalarExpression(ReplaceString)",
             ScalarExpression::Slice(_) => "ScalarExpression(Slice)",
             ScalarExpression::Parse(p) => p.get_name(),
             ScalarExpression::Temporal(t) => t.get_name(),
+            ScalarExpression::Text(r) => r.get_name(),
             ScalarExpression::Math(m) => m.get_name(),
         }
     }
@@ -832,123 +831,6 @@ impl Expression for LengthScalarExpression {
 
     fn get_name(&self) -> &'static str {
         "LengthScalarExpression"
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ReplaceStringScalarExpression {
-    query_location: QueryLocation,
-    haystack_expression: Box<ScalarExpression>,
-    needle_expression: Box<ScalarExpression>,
-    replacement_expression: Box<ScalarExpression>,
-    case_insensitive: bool,
-}
-
-impl ReplaceStringScalarExpression {
-    pub fn new(
-        query_location: QueryLocation,
-        haystack_expression: ScalarExpression,
-        needle_expression: ScalarExpression,
-        replacement_expression: ScalarExpression,
-        case_insensitive: bool,
-    ) -> ReplaceStringScalarExpression {
-        Self {
-            query_location,
-            haystack_expression: haystack_expression.into(),
-            needle_expression: needle_expression.into(),
-            replacement_expression: replacement_expression.into(),
-            case_insensitive,
-        }
-    }
-
-    pub fn get_haystack_expression(&self) -> &ScalarExpression {
-        &self.haystack_expression
-    }
-
-    pub fn get_needle_expression(&self) -> &ScalarExpression {
-        &self.needle_expression
-    }
-
-    pub fn get_replacement_expression(&self) -> &ScalarExpression {
-        &self.replacement_expression
-    }
-
-    pub fn get_case_insensitive(&self) -> bool {
-        self.case_insensitive
-    }
-
-    pub(crate) fn try_resolve_value_type(
-        &self,
-        pipeline: &PipelineExpression,
-    ) -> Result<Option<ValueType>, ExpressionError> {
-        if let Some(v) = self
-            .get_haystack_expression()
-            .try_resolve_value_type(pipeline)?
-        {
-            Ok(Some(match v {
-                ValueType::String => ValueType::String,
-                _ => ValueType::Null,
-            }))
-        } else {
-            Ok(None)
-        }
-    }
-
-    pub(crate) fn try_resolve_static<'a, 'b, 'c>(
-        &'a self,
-        pipeline: &'b PipelineExpression,
-    ) -> Result<Option<ResolvedStaticScalarExpression<'c>>, ExpressionError>
-    where
-        'a: 'c,
-        'b: 'c,
-    {
-        let haystack_static = self
-            .get_haystack_expression()
-            .try_resolve_static(pipeline)?;
-        let needle_static = self.get_needle_expression().try_resolve_static(pipeline)?;
-        let replacement_static = self
-            .get_replacement_expression()
-            .try_resolve_static(pipeline)?;
-
-        if let (Some(haystack), Some(needle), Some(replacement)) =
-            (haystack_static, needle_static, replacement_static)
-        {
-            let haystack_value = haystack.to_value();
-            let needle_value = needle.to_value();
-            let replacement_value = replacement.to_value();
-
-            if let Some(result) = Value::replace_matches(
-                &haystack_value,
-                &needle_value,
-                &replacement_value,
-                self.case_insensitive,
-            ) {
-                Ok(Some(ResolvedStaticScalarExpression::Value(
-                    StaticScalarExpression::String(StringScalarExpression::new(
-                        self.query_location.clone(),
-                        &result,
-                    )),
-                )))
-            } else {
-                Ok(Some(ResolvedStaticScalarExpression::Value(
-                    StaticScalarExpression::Null(NullScalarExpression::new(
-                        self.query_location.clone(),
-                    )),
-                )))
-            }
-        } else {
-            Ok(None)
-        }
-    }
-}
-
-impl Expression for ReplaceStringScalarExpression {
-    fn get_query_location(&self) -> &QueryLocation {
-        &self.query_location
-    }
-
-    fn get_name(&self) -> &'static str {
-        "ReplaceStringScalarExpression"
     }
 }
 
@@ -1805,124 +1687,6 @@ mod tests {
             (
                 ScalarExpression::Static(StaticScalarExpression::Boolean(
                     BooleanScalarExpression::new(QueryLocation::new_fake(), true),
-                )),
-                Some(ValueType::Null),
-                Some(Value::Null),
-            ),
-        ]);
-    }
-
-    #[test]
-    #[allow(clippy::type_complexity)]
-    pub fn test_replace_string_scalar_expression_try_resolve() {
-        fn run_test(
-            input: Vec<(
-                ScalarExpression,
-                ScalarExpression,
-                ScalarExpression,
-                Option<ValueType>,
-                Option<Value>,
-            )>,
-        ) {
-            for (text, lookup, replacement, expected_type, expected_value) in input {
-                let e = ReplaceStringScalarExpression::new(
-                    QueryLocation::new_fake(),
-                    text,
-                    lookup,
-                    replacement,
-                    false, // case_insensitive
-                );
-
-                let pipeline = Default::default();
-
-                let actual_type = e.try_resolve_value_type(&pipeline).unwrap();
-                assert_eq!(expected_type, actual_type);
-
-                let actual_value = e.try_resolve_static(&pipeline).unwrap();
-                assert_eq!(expected_value, actual_value.as_ref().map(|v| v.to_value()));
-            }
-        }
-
-        run_test(vec![
-            (
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(
-                        QueryLocation::new_fake(),
-                        "A magic trick can turn a cat into a dog",
-                    ),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "cat"),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "hamster"),
-                )),
-                Some(ValueType::String),
-                Some(Value::String(&StringScalarExpression::new(
-                    QueryLocation::new_fake(),
-                    "A magic trick can turn a hamster into a dog",
-                ))),
-            ),
-            (
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "hello world hello"),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "hello"),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "hi"),
-                )),
-                Some(ValueType::String),
-                Some(Value::String(&StringScalarExpression::new(
-                    QueryLocation::new_fake(),
-                    "hi world hi",
-                ))),
-            ),
-            (
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "no matches here"),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "xyz"),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "abc"),
-                )),
-                Some(ValueType::String),
-                Some(Value::String(&StringScalarExpression::new(
-                    QueryLocation::new_fake(),
-                    "no matches here",
-                ))),
-            ),
-            (
-                ScalarExpression::Source(SourceScalarExpression::new(
-                    QueryLocation::new_fake(),
-                    ValueAccessor::new_with_selectors(vec![ScalarExpression::Static(
-                        StaticScalarExpression::String(StringScalarExpression::new(
-                            QueryLocation::new_fake(),
-                            "key1",
-                        )),
-                    )]),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "search"),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "replace"),
-                )),
-                None,
-                None,
-            ),
-            (
-                ScalarExpression::Static(StaticScalarExpression::Boolean(
-                    BooleanScalarExpression::new(QueryLocation::new_fake(), true),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "search"),
-                )),
-                ScalarExpression::Static(StaticScalarExpression::String(
-                    StringScalarExpression::new(QueryLocation::new_fake(), "replace"),
                 )),
                 Some(ValueType::Null),
                 Some(Value::Null),
